requirements.txt
django
djangorestframework
grpcio
grpcio-tools
